requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
pytest-django==4.7.0
pytest==7.4.3
python-decouple==3.8
